moving files out of core folder
